moved input logic from UI to Game
