+++ src/game.py
@@ -1,3 +1,5 @@
+import re
+
 from src.board import Board
 from src.message import Message
 from src.rules import Rules
@@ -25,6 +27,42 @@ class Game:
         else:
             return self.player_two
 
+    def validate_move(self):
+        user_input = self.ui.get_user_input()
+        is_integer = self.is_integer(user_input)
+        if is_integer:
+            user_integer = int(user_input)
+            return self.validate_move_for_range(user_integer)
+        else:
+            return self.handle_incorrect_input(user_input)
+
+    def validate_move_for_range(self, user_input):
+        if self.input_in_range(user_input):
+            return user_input
+        else:
+            return self.handle_incorrect_input(user_input)
+
+    def input_in_range(self, user_input):
+        if user_input in range(1, 10):
+            return True
+        else:
+            return False
+
+    def is_integer(self, user_input):
+        if re.match(r"^[0-9]*$", user_input) is None:
+            return False
+        else:
+            return True
+
+    def handle_incorrect_input(self, user_input):
+        valid_move = self.input_in_range(user_input)
+        while valid_move is False:
+            self.ui.display_message(self.message.incorrect_board_input())
+            user_input = self.validate_move()
+            valid_move = self.input_in_range(user_input)
+            break
+        return user_input
+
     def get_prompt(self, total_marks_on_board):
         current_player = self.get_current_player(total_marks_on_board)
         if self.board.is_full(total_marks_on_board, self.game_board):
@@ -33,7 +71,7 @@ class Game:
             self.ui.display_message(self.message.prompt_for_move(current_player))
 
     def process_user_input(self):
-        position_choice = self.ui.get_user_input(self.message.incorrect_board_input())
+        position_choice = self.validate_move()
         if self.board.determine_is_spot_taken(self.game_board, position_choice):
             self.ui.display_message(self.message.spot_taken_message())
             self.process_user_input()
