+++ src/user_interface.py
@@ -1,6 +1,3 @@
-import re
-
-
 class UserInterface:
     def display_message(self, message):
         print(message)
@@ -8,40 +5,8 @@ class UserInterface:
     def display_board(self, board):
         print(board)
 
-    def get_user_input(self, message):
+    def get_user_input(self):
         user_input = input()
-        is_integer = self.is_integer(user_input)
-        if is_integer:
-            user_integer = int(user_input)
-            return self.validate_move(user_integer, message)
-        else:
-            return self.handle_incorrect_input(user_input, message)
-
-    def validate_move(self, user_input, message):
-        if self.input_in_range(user_input):
-            return user_input
-        else:
-            return self.handle_incorrect_input(user_input, message)
-
-    def input_in_range(self, user_input):
-        if user_input in range(1, 10):
-            return True
-        else:
-            return False
-
-    def is_integer(self, user_input):
-        if re.match(r"^[0-9]*$", user_input) is None:
-            return False
-        else:
-            return True
-
-    def handle_incorrect_input(self, user_input, message):
-        valid_move = self.input_in_range(user_input)
-        while valid_move is False:
-            self.display_message(message)
-            user_input = self.get_user_input(message)
-            valid_move = self.input_in_range(user_input)
-            break
         return user_input
 
     def get_play_again_user_input(self, message):
